chore: convert index.js to TypeScript
